Extract calculateHours
